==> 09_data_prep_additional_columns.ipynb <==
# Tutorial 9: Data Preparation to create additional columns with fake data

In this notebook, we will demonstrate how to create additional columns using [Faker](https://faker.readthedocs.io/en/master/index.html)

## Steps:

- Load _creditcard.csv_ into a pandas dataframe
- Use faker to generate additional columns corresponding to dates, ip addresses, longitude, latitude etc.
- Use sklearn **train_test_split** function along with **stratify** option to split this imbalanced data into test and train data sets
- Save the data frames into two separate files: _credit_card_train_new_columns.csv_ , _credit_card_test_new_columns.csv_

import pandas as pd
from sklearn.model_selection   import  train_test_split
from faker import Faker
from tqdm import tqdm
import random

full_path = 'creditcard.csv'
# load the dataset
data = pd.read_csv(full_path, header=None)
y = data.values[:,-1]

column_names = ["V"+str(i) for i in range(30)]
column_names.append("Target")
data.columns = column_names

Faker.seed(0)
fake = Faker()

## Generate time pre and time post

rows_missing_time_variable = random.sample(range(0,len(data)),100)

time_pre_list = list()
time_post_list = list()
for i in tqdm(range(0,len(data))):
    if i not in rows_missing_time_variable:
        time_pre  = fake.date_time_between(start_date='-2w',end_date='-1w')
        time_post = fake.date_time_between(start_date='-1w')
    else:
        time_pre = None
        time_post = None
    time_pre_list.append(time_pre)
    time_post_list.append(time_post)

data['time_pre'] = time_pre_list
data['time_post'] = time_post_list

## Generate credit card expiry date

rows_missing_cc_expiry_date = random.sample(range(0,len(data)),150)

cc_exp_list = list()
for i in tqdm(range(0,len(data))):
    if i not in rows_missing_cc_expiry_date:
        cc_exp = fake.credit_card_expire()
    else:
        cc_exp = None
    cc_exp_list.append(cc_exp)

data['cc_exp'] = cc_exp_list

## Generate three lat long

rows_missing_lat_long = random.sample(range(0,len(data)),200)

lat1_list = list()
lng1_list = list()
lat2_list = list()
lng2_list = list()
lat3_list = list()
lng3_list = list()
for i in tqdm(range(0,len(data))):
    if i not in rows_missing_lat_long:
        lat1, lng1 = fake.latlng()
        lat2, lng2 = fake.latlng()
        lat3, lng3 = fake.latlng()
    else:
        lat1, lng1, lat2, lng2, lat3, lng3 = [None]*6
    lat1_list.append(lat1)
    lng1_list.append(lng1)
    lat2_list.append(lat2)
    lng2_list.append(lng2)
    lat3_list.append(lat3)
    lng3_list.append(lng3)

data['lat1'] = lat1_list
data['long1'] = lng1_list
data['lat2'] = lat2_list
data['long2'] = lng2_list
data['lat3'] = lat3_list
data['long3'] = lng3_list

## Generate IP addresses

rows_missing_ip_addresses = random.sample(range(0,len(data)),120)

ip_list = list()
for i in tqdm(range(0,len(data))):
    if i not in rows_missing_ip_addresses:
        ipv4 = fake.ipv4_public()
    else:
        ipv4 = None
    ip_list.append(ipv4)

data['ip_address'] = ip_list

data

df_train,df_test = train_test_split(data,test_size=0.2,stratify=y,random_state=42)

df_train.to_csv("credit_card_train_new_columns.csv",index=False)
df_test.to_csv("credit_card_test_new_columns.csv",index=False)
