==> 11_model_rest_api_inference_filter_transforms.ipynb <==
# Tutorial 11: Test inference + shapley values REST API Endpoint 

In this notebook, we will demonstrate how to perform inference using an ML model deployed as a REST API endpoint. REST API will also filter columns that are not needed for the prediction. Output of the inference will include  reason codes (shapley values) in addition to the prediction.

## Steps
- Import requests and pandas libraries. Requests is an elegant and simple HTTP library for Python, built for human beings. More about requests [here](https://docs.python-requests.org/en/latest/)
- Create a pandas dataframe for the data on which we want to perform inference
- Copy the REST API endpoint from _Models_ tab
- Create a json from the pandas dataframe
- Add additional columns to test column filter
- Create a header for a post request
- Perform a post request on the REST API endpoint

import requests
import pandas as pd
import numpy as np
import joblib

## Get test data

input_df = pd.read_csv("credit_card_test_new_columns.csv")
y_test = input_df["Target"]
X_test = input_df.drop(["Target"],axis=1)

## Add extra columns to reproduce real case

X_test["V100"] = X_test["V22"]
X_test["V200"] = X_test["V23"]*0.9

## Test REST API Endpoint

%%time 

url = "https://k1yltn5.rmlpoc.net/invocations"
input_data = X_test.head(10)
data_json = input_data.to_json(orient="split",index=False)
headers = {"Content-Type":"application/json; format=pandas-split"}
response = requests.post(url,data=data_json,headers=headers)

response.json()

input_data
