==> 14_Model Endpoint_xgboost.ipynb <==
# Tutorial 13: Xgboost Model Testing

In this notebook, we will demonstrate how to perform inference using an ML model deployed as a REST API endpoint. This model has been created using Tutorial 12

## Steps
- Import requests and pandas libraries. Requests is an elegant and simple HTTP library for Python, built for human beings. More about requests here
- Create a pandas dataframe for the data on which we want to perform inference
- Copy the REST API endpoint from Models tab (i.e. model has been converted as docker container)
- Create a json from the pandas dataframe
- Create a header for a post request
- Perform a post request on the REST API endpoint

import requests
import pandas as pd
import numpy as np

### Get the required data

# Download the data set
from sklearn.model_selection import train_test_split
from sklearn.datasets import load_diabetes
from sklearn.tree import DecisionTreeRegressor

db = load_diabetes()
db_df_Ind = pd.DataFrame(db.data,columns=["db1","db2","db3","db4","db5","db6","db7","db8","db9","db10"])
db_df_Target = pd.DataFrame(db.target,columns=["Target"])
print(db_df_Ind.columns.values)
print(db_df_Target.columns.values)

#### Split the data into train and test

X_train, X_test, y_train, y_test = train_test_split(db_df_Ind, db_df_Target,
                                                    test_size=0.20,
                                                    random_state=10)

#### Send HTTP post request to the model end point

url = "https://wfav8nq.ostack.babyrocket.net/invocations"
data_json = X_test.head(5).to_json(orient="split",index=False)
headers = {"Content-Type":"application/json; format=pandas-split"}
response = requests.post(url,data=data_json,headers=headers)

# Print the response
response.json()
